changed cluster View,icons on menu and colors on status bar

+++ app/src/main/java/com/example/hataru/presentation/forMap/GeometryProvider.kt
@@ -21,8 +21,8 @@ object GeometryProvider {
     }
 
 
-    private fun getRandomCost() : Double{
-        return (1000..3000).random().toDouble()
+    private fun getRandomCost() : Int{
+        return (1000..3000).random()
     }
 
     val clusterizedPoints = listOf(
