+++ app/src/main/java/com/example/hataru/presentation/forMap/flat.kt
@@ -2,4 +2,4 @@ package com.example.hataru.presentation.forMap
 
 import com.yandex.mapkit.geometry.Point
 
-data class flat(val id : Int,val location : Point,val cost : Double){}
+data class flat(val id : Int,val location : Point,val cost : Int){}
